+++ inc/includes.php
@@ -48,9 +48,6 @@ if (!defined('GLPI_ROOT')) {
 
 include_once GLPI_ROOT . '/inc/based_config.php';
 
-// Init Timer to compute time of display
-$TIMER_DEBUG = new Timer();
-$TIMER_DEBUG->start();
 \Glpi\Debug\Profiler::getInstance()->start('php_request');
 
 
@@ -77,12 +74,6 @@ if (
 ) {
     // Start the debug profile
     $profile = \Glpi\Debug\Profile::getCurrent();
-    $SQL_TOTAL_REQUEST    = 0;
-    $DEBUG_SQL = [
-        'queries' => [],
-        'errors'  => [],
-        'times'   => [],
-    ];
 }
 
 // Mark if Header is loaded or not :