+++ src/Console/Application.php
@@ -310,25 +310,12 @@ class Application extends BaseApplication
     /**
      * Initalize GLPI.
      *
-     * @global array $CFG_GLPI
      * @global GLPI  $GLPI
      *
      * @return void
      */
     private function initApplication()
     {
-
-       // Disable debug at bootstrap (will be re-enabled later if requested by verbosity level).
-        /** @var array $CFG_GLPI */
-        global $CFG_GLPI;
-        $CFG_GLPI = array_merge(
-            $CFG_GLPI,
-            [
-                'debug_sql'  => 0,
-                'debug_vars' => 0,
-            ]
-        );
-
         /** @var \GLPI $GLPI */
         global $GLPI;
         $GLPI = new GLPI();